Refactor tests for the future implementation of integration tests
