+++ package.json
@@ -7,7 +7,7 @@
     "test": "test"
   },
   "scripts": {
-    "test": "NODE_ENV=test istanbul cover ./node_modules/mocha/bin/_mocha --report lcov -- -R spec --timeout 10000 && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
+    "test": "NODE_ENV=test istanbul cover ./node_modules/mocha/bin/_mocha --report lcov -- -R spec --timeout 10000 test/**/*.js && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
     "lint": "node_modules/.bin/eslint lib/"
   },
   "repository": {
